clean up

+++ composable_kernel/include/utility/magic_division.hpp
@@ -141,7 +141,7 @@ struct MagicDivision
     DoMagicDivision(int32_t dividend_i32, uint32_t multiplier, uint32_t shift)
     {
         uint32_t dividend_u32 = bit_cast<uint32_t>(dividend_i32);
-        uint32_t tmp          = static_cast<unsigned long long>(dividend_u32) * multiplier >> 32;
+        uint32_t tmp          = static_cast<uint64_t>(dividend_u32) * multiplier >> 32;
         return (tmp + dividend_u32) >> shift;
     }
 };
